Moved non-ROCm related packages to /opt

+++ extras/scripts/parmetis_setup.sh
@@ -5,11 +5,12 @@ AMDGPU_GFXMODEL=`rocminfo | grep gfx | sed -e 's/Name://' | head -1 |sed 's/ //g
 MODULE_PATH=/etc/lmod/modules/misc/parmetis
 BUILD_PARMETIS=1
 ROCM_VERSION=6.4.0
-INSTALL_PATH=/opt/rocmplus-${ROCM_VERSION}/parmetis-v3.14
+PARMETIS_VERSION="3.14"
+INSTALL_PATH=/opt/parmetis-v${PARMETIS_VERSION}
 INSTALL_PATH_INPUT=""
 SUDO="sudo"
-MPI_MODULE="openmpi"
 DEB_FRONTEND="DEBIAN_FRONTEND=noninteractive"
+MPI_MODULE="openmpi"
 
 if [  -f /.singularity.d/Singularity ]; then
    SUDO=""
@@ -27,8 +28,9 @@ usage()
    echo "  WARNING: when selecting the module to supply to --mpi-module, make sure it sets the MPI_PATH environment variable"
    echo "  --module-path [ MODULE_PATH ] default $MODULE_PATH"
    echo "  --rocm-version [ ROCM_VERSION ] default $ROCM_VERSION"
-   echo "  --install-path [ INSTALL_PATH_INPUT ] default $INSTALL_PATH"
    echo "  --mpi-module [ MPI_MODULE ] default $MPI_MODULE"
+   echo "  --install-path [ INSTALL_PATH_INPUT ] default $INSTALL_PATH"
+   echo "  --parmetis_version [ PARMETIS_VERSION ] default $PARMETIS_VERSION"
    echo "  --amdgpu-gfxmodel [ AMDGPU-GFXMODEL ] default autodetected"
    echo "  --build-parmetis [ BUILD_PARMETIS ] default is 0"
    echo "  --help: this usage information"
@@ -69,6 +71,11 @@ do
           MODULE_PATH=${1}
           reset-last
           ;;
+      "--parmetis-version")
+          shift
+          PARMETIS_VERSION=${1}
+          reset-last
+          ;;
       "--mpi-module")
           shift
           MPI_MODULE=${1}
@@ -98,8 +105,8 @@ done
 if [ "${INSTALL_PATH_INPUT}" != "" ]; then
    INSTALL_PATH=${INSTALL_PATH_INPUT}
 else
-   # override path in case ROCM_VERSION has been supplied as input
-   INSTALL_PATH=/opt/rocmplus-${ROCM_VERSION}/parmetis
+   # override path in case PARMETIS_VERSION has been supplied as input
+   INSTALL_PATH=/opt/parmetis-v${PARMETIS_VERSION}
 fi
 
 echo ""
@@ -109,7 +116,8 @@ echo "ROCM_VERSION: $ROCM_VERSION"
 echo "BUILD_PARMETIS: $BUILD_PARMETIS"
 echo "Installing PARMETIS in: $INSTALL_PATH"
 echo "MODULE_PATH: $MODULE_PATH"
-echo "Loading this module for MPI: $MPI_MODULE"
+echo "PARMETIS_VERSION: $PARMETIS_VERSION"
+echo "MPI_MODULE: $MPI_MODULE"
 echo "==================================="
 echo ""
 
@@ -131,7 +139,7 @@ else
       echo ""
 
       #install the cached version
-      cd /opt/rocmplus-${ROCM_VERSION}
+      cd /opt
       tar -xpzf ${CACHE_FILES}/parmetis.tgz
       if [ "${USER}" != "sysadmin" ]; then
          ${SUDO} rm ${CACHE_FILES}/parmetis.tgz
@@ -153,7 +161,11 @@ else
       source /etc/profile.d/lmod.sh
       source /etc/profile.d/z01_lmod.sh
       module load gcc
-      module load openmpi
+      module load $MPI_MODULE
+      if [[ $MPI_PATH == "" ]]; then
+         echo "MPI module $MPI_MODULE is not setting the MPI_PATH env variable, aborting..."
+         exit 1
+      fi
 
       ${SUDO} rm -rf parmetis
       ${SUDO} rm -rf $INSTALL_PATH
@@ -184,7 +196,7 @@ else
       cd ..
 
       cd ..
-      rm -rf parmetis
+      rm -rf gklib metis parmetis
       rm -f gklib_force_fpic.patch
 
       if [[ "${USER}" != "root" ]]; then
@@ -195,15 +207,27 @@ else
       fi
    fi
 
+   # Create a module file for fftw
+   if [ -d "$MODULE_PATH" ]; then
+      # use sudo if user does not have write access to module path
+      if [ ! -w ${MODULE_PATH} ]; then
+         SUDO="sudo"
+      else
+         echo "WARNING: not using sudo since user has write access to module path"
+      fi
+   else
+      # if module path dir does not exist yet, the check on write access will fail
+      SUDO="sudo"
+      echo "WARNING: using sudo, make sure you have sudo privileges"
+   fi
+
    ${SUDO} mkdir -p ${MODULE_PATH}
 
-   PARMETIS_PATH=${INSTALL_PATH}
-
    # The - option suppresses tabs
-   cat <<-EOF | ${SUDO} tee ${MODULE_PATH}/dev.lua
+   cat <<-EOF | ${SUDO} tee ${MODULE_PATH}/${PARMETIS_VERSION}.lua
         whatis("PARMETIS package")
 
-        local base = "${PARMETIS_PATH}"
+        local base = "${INSTALL_PATH}"
 
         setenv("PARMETIS", base)
         setenv("PARMETIS_PATH", base)
